Initial refactor for CSV to become a CSV Module
