+++ src/ifcblenderexport/blenderbim/bim/__init__.py
@@ -22,6 +22,7 @@ if bpy is not None:
     import blenderbim.bim.module.type as module_type
     import blenderbim.bim.module.unit as module_unit
     import blenderbim.bim.module.cobie as module_cobie
+    import blenderbim.bim.module.csv as module_csv
     from . import ui, prop, operator
 
     classes = [
@@ -134,11 +135,6 @@ if bpy is not None:
         operator.RemovePropertyTemplate,
         operator.AddSectionPlane,
         operator.RemoveSectionPlane,
-        operator.AddCsvAttribute,
-        operator.RemoveCsvAttribute,
-        operator.ExportIfcCsv,
-        operator.ImportIfcCsv,
-        operator.EyedropIfcCsv,
         operator.ReloadIfcFile,
         operator.AddIfcFile,
         operator.RemoveIfcFile,
@@ -249,7 +245,6 @@ if bpy is not None:
         ui.BIM_PT_document_information,
         ui.BIM_PT_constraints,
         ui.BIM_PT_search,
-        ui.BIM_PT_ifccsv,
         ui.BIM_PT_ifcclash,
         ui.BIM_PT_qa,
         ui.BIM_PT_library,
@@ -302,6 +297,7 @@ if bpy is not None:
     classes.extend(module_type.classes)
     classes.extend(module_unit.classes)
     classes.extend(module_cobie.classes)
+    classes.extend(module_csv.classes)
     
     def menu_func_export(self, context):
         self.layout.operator(operator.ExportIFC.bl_idname, text="Industry Foundation Classes (.ifc/.ifczip/.ifcjson)")
@@ -349,6 +345,7 @@ if bpy is not None:
         module_type.register()
         module_unit.register()
         module_cobie.register()
+        module_csv.register()
         bpy.app.handlers.depsgraph_update_pre.append(operator.depsgraph_update_pre_handler)
         bpy.app.handlers.load_post.append(prop.toggleDecorationsOnLoad)
 
@@ -369,6 +366,7 @@ if bpy is not None:
         del bpy.types.Camera.BIMCameraProperties
         del bpy.types.TextCurve.BIMTextProperties
         bpy.types.SCENE_PT_unit.remove(ui.ifc_units)
+        module_csv.unregister()
         module_cobie.unregister()
         module_unit.unregister()
         module_type.unregister()
